+++ CHANGELOG.md
@@ -21,7 +21,7 @@
 - All builtin bindings support renaming almost all of the interface (types, args, items, variants, etc) via TOML definitions -
   [see the docs](https://mozilla.github.io/uniffi-rs/next/renaming.html).
   ([#2715](https://github.com/mozilla/uniffi-rs/pull/2715))
-- Support for methods on dataclasses for Python. ([#2706](https://github.com/mozilla/uniffi-rs/pull/2706)).
+- Support for methods on records ([#2706](https://github.com/mozilla/uniffi-rs/pull/2706), [#2724](https://github.com/mozilla/uniffi-rs/pull/2724)).
 
 ### What's Fixed